==> 08_pricing_de_ois.ipynb <==
# Valorización de OIS

Entre sí, los bancos operan los instrumentos estándar. En el caso de los OIS sobre SOFR, éstos son instrumentos con periodicidad anual, amortización bullet y a los plazos predefinidos.

¿Qué se hace cuándo es necesario cotizar un swap con características distintas? En esos casos, se hace *pricing* de este contrato especial, utilizando la curva cupón cero que se ha obtenido de los swaps estándar.

## Configuración

### Librerías

from finrisk import QC_Financial_3 as Qcf
from scipy.optimize import root_scalar
import modules.auxiliary as aux
from functools import partial
from enum import Enum
import pandas as pd

### Variables Globales

Formatos para campos de los `DataFrames`.

frmt = {
    'tasa': '{:.6%}',
    'df': '{:.12%}',
    'valor_tasa': '{:.4%}',
    'spread': '{:.4%}',
    'nominal': '{:,.2f}',
    'interes': '{:,.2f}',
    'amortizacion': '{:,.2f}',
    'flujo': '{:,.2f}',
}

## Carga Curva Cero Cupón

Se importa la data de la curva cupón cero que fue construida en el notebook 5.

df_curva = pd.read_excel('data/20201012_built_sofr_zero.xlsx')

df_curva.head(20).style.format(frmt)

zcc = aux.get_curve_from_dataframe(
    Qcf.QCAct365(),
    Qcf.QCCompoundWf(),
    df_curva
)

Algunos métodos del objeto`zcc`.

plazo = 900
print(f"Tasa a {plazo} días es igual a {zcc.get_rate_at(plazo):.4%}")
print(f"Factor de descuento a {plazo} días es igual a {zcc.get_discount_factor_at(plazo):.6%}")

## Pricing

get_ois_sofr = partial(
    aux.get_ois_using_template,
    aux.type_ois_template,
    aux.TypeOis.SOFR
)

### Operación con Plazo Distinto

op = get_ois_sofr(
    rp=Qcf.RecPay.RECEIVE,
    notional=10000000,
    start_date=Qcf.QCDate(14, 10, 2020),
    tenor=Qcf.Tenor('2Y6M'),
    fixed_rate_value=.01,
    spread=0.0,
    gearing=1.0
)
op

aux.show_leg(op[0], 'FixedRateCashflow', '').style.format(frmt)

aux.show_leg(op[1], 'IcpClpCashflow', '').style.format(frmt)

#### Valor Presente Pata Fija

vp = Qcf.PresentValue()

fecha_val = Qcf.QCDate(14, 10, 2020)

vp_fija = vp.pv(fecha_val, op[0], zcc)
print(f'El valor presente de la pata fija es: USD {vp_fija:,.2f}')

#### Valor Presente Pata Flotante

fwd = Qcf.ForwardRates()

fwd.set_rates_icp_clp_leg(fecha_val, 1.0, op[1], zcc)

aux.show_leg(op[1], 'IcpClpCashflow', '').style.format(frmt)

vp_flot = vp.pv(fecha_val, op[1], zcc)
print(f'El valor presente de la pata flotante es: USD {vp_flot:,.2f}')

print(f'Por lo tanto, el valor total de la operación es:\nValor total: USD {vp_fija + vp_flot:,.2f}')

Ecuación a resolver

$$
nominal=c_{1Y}\cdot df_{1Y}+c_{2Y}\cdot df_{2Y}
$$

$$
nominal = \left[r\cdot yf_1 \cdot df_{1Y}+ \left(1+r\cdot yf_2\right) \cdot df_{2Y}\right]\cdot nominal
$$

$$
\frac{1 - df_{1Y}}{\cdot\left(yf_1\cdot df_{1Y}+yf_2\cdot df_{2Y}\right)}=r
$$

#### Ejercicio

Haga el *pricing* de la operación: determine qué tasa fija hace que el valor total de la operación sea 0.

get_ois_sofr_solo_tasa = partial(
    aux.get_ois_using_template,
    template=aux.type_ois_template,
    rp=Qcf.RecPay.RECEIVE,
    type_ois=aux.TypeOis.SOFR,
    notional=10000000,
    start_date=Qcf.QCDate(14, 10, 2020),
    tenor=Qcf.Tenor('2Y6M'),
    spread=0.0,
    gearing=1.0
)

get_ois_sofr_solo_tasa(fixed_rate_value=0.0)

def error(fixed_rate_value: float) -> float:
    this_op = get_ois_sofr_solo_tasa(fixed_rate_value=fixed_rate_value)
    err = vp.pv(fecha_val, this_op[0], zcc) - 10000000
    return err

tasa = .01
print(f'Error: {error(tasa):,.2f}')

x = root_scalar(
        error,
        method='bisect',
        bracket=[0.0, .02],
        x0=.01,
        xtol=.00000000000000001
)

print(f'Tasa fija es: {x.root:.8%}')

Dejo a ustedes comprobar la solución.

### Operación con Amortizaciones

op2 = get_ois_sofr(
    rp=Qcf.RecPay.RECEIVE,
    notional=10000000,
    start_date=Qcf.QCDate(14, 10, 2020),
    tenor=Qcf.Tenor('4Y'),
    fixed_rate_value=.01,
    spread=0.0,
    gearing=1.0
)
op2

aux.show_leg(op2[0], 'FixedRateCashflow', '').style.format(frmt)

aux.show_leg(op2[1], 'IcpClpCashflow', '').style.format(frmt)

#### Agrega Amortizaciones

Vamos a agregar una amortización por la mitad del nocional en el segundo cupón de cada pata. Veamos primero la pata fija.

cshflw = op2[0].get_cashflow_at(1)
cshflw.set_amortization(5000000)

aux.show_leg(op2[0], 'FixedRateCashflow', '').style.format(frmt)

Vemos que la amortización queda bien ingresada, sin embargo, los flujos siguientes no la consideran. Vamos a arreglar eso:

cshflw = op2[0].get_cashflow_at(2)
cshflw.set_nominal(5000000)

cshflw = op2[0].get_cashflow_at(3)
cshflw.set_nominal(5000000)
cshflw.set_amortization(5000000)

aux.show_leg(op2[0], 'FixedRateCashflow', '').style.format(frmt)

Hagamos ahora la pata flotante. Dado que esta es la pata que pagamos, hay que usar signo negativo para los montos de amortización y capital vigente.

cshflw = op2[1].get_cashflow_at(1)
cshflw.set_amortization(-5000000)

cshflw = op2[1].get_cashflow_at(2)
cshflw.set_nominal(-5000000)

cshflw = op2[1].get_cashflow_at(3)
cshflw.set_nominal(-5000000)
cshflw.set_amortization(-5000000)

aux.show_leg(op2[1], 'IcpClpCashflow', '').style.format(frmt)

#### Valor Presente por Pata

Calculemos ahora los valores presente de cada pata.

vp_fija_2 = vp.pv(fecha_val, op2[0], zcc)
print(f'El valor presente de la pata fija es: USD {vp_fija_2:,.2f}')

fwd.set_rates_icp_clp_leg(fecha_val, 1.0, op2[1], zcc)
vp_flot_2 = vp.pv(fecha_val, op2[1], zcc)
print(f'El valor presente de la pata flotante es: USD {vp_flot_2:,.2f}')

print(f'Por lo tanto, el valor total de la operación es:\nValor total: USD {vp_fija_2 + vp_flot_2:,.2f}')

#### Ejercicio

Pricee esta operación.

Segunda forma de obtener una función de 1 sola variable. Se define una función que **retorna** una función de 1 sola variable (que representa el valor de la tasa fija de la operación).

def get_sofr_ois_builder_only_rate(
    rp,
    notional,
    start_date,
    tenor,
    spread,
    gearing
):
    def f(fixed_rate_value):
        return get_ois_sofr(
            rp,
            notional,
            start_date,
            tenor,
            fixed_rate_value,
            spread,
            gearing
        )
    return f

Utilizamos la función anterior y obtenemos una función que construye un OIS SOFR a partir del valor de la tasa fija.

get_ois_sofr_solo_tasa2 = get_sofr_ois_builder_only_rate(
    rp=Qcf.RecPay.RECEIVE,
    notional=10000000,
    start_date=Qcf.QCDate(14, 10, 2020),
    tenor=Qcf.Tenor('4Y'),
    spread=0.0,
    gearing=1.0
)

test = get_ois_sofr_solo_tasa2(.03)

aux.show_leg(test[0], 'FixedRateCashflow', '').style.format(frmt)

Con la ayuda de la función anterior se obtiene una función `error2` que será la que utilizaremos para el *pricing*.

def error2(fixed_rate_value: float) -> float:
    this_op = get_ois_sofr_solo_tasa2(fixed_rate_value=fixed_rate_value)
    
    cshflw = this_op[0].get_cashflow_at(1)
    cshflw.set_amortization(5000000)
    
    cshflw = this_op[0].get_cashflow_at(2)
    cshflw.set_nominal(5000000)

    cshflw = this_op[0].get_cashflow_at(3)
    cshflw.set_nominal(5000000)
    cshflw.set_amortization(5000000)
    
    err = vp.pv(fecha_val, this_op[0], zcc) - 10000000 # Esto es muy ordinario.
    # En el caso general en vez de 10000000 habría que calcular al valor presente
    # de la pata OIS.
    
    return err

Se prueba la función `error2`.

tasa = .01
print(f'Error: {error2(tasa):,.2f}')

Se hace el *pricing*.

x2 = root_scalar(
        error2,
        method='bisect',
        bracket=[0.0, .02],
        x0=.01,
        xtol=.00000000000000001
)

print(f'Tasa fija es: {x2.root:.8%}')

##### Ejercicio

Implementar las modificaciones necesarias para poder hacer el *pricing* de una operación cuya pata OIS tenga spread.
