==> 20_Tarea_1.ipynb <==
# Tarea 1

import pandas as pd

## Tarea

Replicar en Excel el cálculo de la curva cupón cero en el notebook **07_sensibilidad_ois.ipynb**. Recordar que las tasas están en convención Exp Act/365.

**Nota:** no sólo se evaluará el resultado obtenido con la planilla, se dará mucha importancia, también, a la arquitectura de la solución. Se recomienda pensar que la planilla se transformará en una herramienta que se utilizará diariamente en un entorno real.

Los valores de la curva obtenida son:

df_curva = pd.read_excel('data/20201012_built_sofr_zero.xlsx')

df_curva.style.format({'tasa': '{:.4%}', 'df': '{:.6%}'})
