==> 22_examen_final.ipynb <==
# Examen Final

- Prender cámaras.
- Puede utilizar sus apuntes.
- Puede utilizar los notebooks del curso.
- La alumna o alumno que sea sorprendido solicitando ayuda a cualquier persona, sea ella compañera o no, será calificada con un 2.
- Cada pregunta vale un punto.
- La nota mínima es un 2.
- La nota se determinará con la siguiente fórmula $nota = \min\left(puntos + 2, 7\right)$
- Todas las funciones necesarias para resolver los ejercicios están ya importadas al notebook.

## Librerías

from finrisk import examen_final as ef
from scipy.interpolate import interp1d
import modules.hull_white as hw
import pandas as pd
import numpy as np
import textwrap
import random
import math

## Curva Cero Cupón

Los plazos están en días y las tasas en convención exp act/365.

df_curva = pd.read_excel('data/20201012_built_sofr_zero.xlsx')

df_curva.head()

curva = interp1d(df_curva['plazo'], df_curva['tasa'], 'linear', fill_value='extrapolate')

## Cálculo Opción

Para identificar si es una Call o una Put se usa un `enum`.

c_p = hw.CallPut.CALL
c_p = hw.CallPut.PUT

Ver toda la documentación:

print(hw.zcb_call_put.__doc__)

## Valores `gamma` y `sigma`

gamma = 1
sigma = .005

## Preguntas

try:
    q = ef.get_questions()
except Exception as e:
    print(str(e))

for qq in q:
    print(f'{qq[0]}\n')
    print(textwrap.fill(f'{qq[1]}', 80))
    print('\n')

## Respuestas

Responder en el siguiente espacio. Por favor incluir explicaciones y comentarios para facilitar la labor de quien corrige.

### Pregunta 1

### Pregunta 2

### Pregunta 3

### Pregunta 4

### Pregunta 5

### Pregunta 6

### Pregunta 7
