==> 23_pauta_tarea_2.ipynb <==
# Pauta Tarea 2

## Librerías

import math
import numpy as np
import pandas as pd
from scipy.stats import norm
from dataclasses import dataclass
from typing import Iterable, Tuple
from scipy.optimize import minimize

import modules.hull_white as hw

## Importa Data

data = pd.read_excel(
    'data_enunciado_tarea_2.xlsx',
    sheet_name='Sheet1',
    usecols='A:D',
)
data.columns = ['maturity', 'tasa_swap', 'df', 'cap_vol']

Formato para el `DataFrame`.

frmt = {
    'maturity': '{:.2f}',
    'tasa_swap': '{:.4%}',
    'df': '{:.8%}',
    'cap_vol': '{:.5%}'
}

Se visualiza.

data.style.format(frmt)

## Modelo para un Cap y sus Caplets

### Clase `Caplet`

@dataclass
class Caplet:
    notional: float
    start_time: float
    end_time: float
    strike: float
        
    def get_yf(self):
        return self.end_time - self.start_time

### Alias `Cap`

Cap = Iterable[Caplet]

### Factory Function

def make_cap(
    strike: float,
    num_caplets: int,
    periodicity: float,
    notional: float
) -> Cap:
    """
    Construye un Cap.
    
    - strike: strike del Cap.
    - num_caplets: número de caplets. Debe incluir el caplet que comienza en `t = 0`.
    - periodicity: periodicidad de los caplets (equivalente al tenor del índice). Se expresa en fracción
    de año canónica, por ejemplo 3M -> 0.25, 6M -> .50.
    - notional: nocional del CAP.
    """
    result = []
    for i in range(num_caplets):
        start_time = i * periodicity
        end_time = (i + 1) * periodicity
        result.append(
            Caplet(
                notional=notional,
                start_time=start_time,
                end_time=end_time,
                strike=strike
            )
        )
    return result

test_cap = make_cap(
    strike=.026,
    num_caplets=4,
    periodicity=.25,
    notional=10000000
)

test_cap

## Modelo para la Curva Cero

Se define una clase que contenga los factores de descuento de la curva cero cupón. La clase define un método de interpolación log-lineal en dichos factores de descuento.

@dataclass
class DiscountFactorCurve:
    tenors_dfs: Iterable[Tuple[float, float]]

    def __post_init__(self):
        self.tenors_dfs.sort(key=lambda x: x[0])

    def get_df_at(self, t: float) -> float:
        return math.exp(
            np.interp(
                t,
                xp=[x[0] for x in self.tenors_dfs],
                fp=[math.log(x[1]) for x in self.tenors_dfs],
            ))

    def get_linear_rate_at(self, t: float) -> float:
        df = self.get_df_at(t)
        return (1 / df - 1) / t
    
    def get_continous_rate_at(self, t: float) -> float:
        df = self.get_df_at(t)
        return -math.log(df) / t
        
    def get_df_fwd_between(self, t0: float, t1: float) -> float:
        return self.get_df_at(t1) / self.get_df_at(t0)

    def get_linear_fwd_rate_between(self, t0: float, t1: float) -> float:
        df_fwd = self.get_df_fwd_between(t0, t1)
        return (1 / df_fwd - 1) / (t1 - t0)

Se obtiene la data y se da de alta el objeto.

tenors_dfs = [(x[0], x[2]) for x in data.values]

tenors_dfs

### Se Prueban los Métodos

No se trata de test unitarios tradicionales ya que los resultados no se comparan con resultados esperados.

df_curve = DiscountFactorCurve(tenors_dfs)
print(df_curve.get_df_at(0.0))
print(df_curve.get_linear_rate_at(.25))
print(df_curve.get_df_fwd_between(0.0, .25))
print(df_curve.get_linear_fwd_rate_between(1.0, 2.0))

## Funciones de Valorización

Las siguientes funciones permiten valorizar un caplet, y en consecuencia un cap, usando la volatilidad de mercado y la fórmula de Black-Scholes-Merton.

def value_caplet(
    caplet: Caplet,
    dfs: DiscountFactorCurve,
    vol: float,
) -> float:
    if caplet.start_time < 0.0:
        return 0.0
    elif caplet.start_time == 0.0:
        yf = caplet.get_yf()
        fwd_rate = dfs.get_linear_fwd_rate_between(
            caplet.start_time,
            caplet.end_time
        )
        return caplet.notional * yf * max(fwd_rate - caplet.strike, 0.0)
    else:
        yf = caplet.get_yf()
        fwd_rate = dfs.get_linear_fwd_rate_between(
            caplet.start_time,
            caplet.end_time
        )
        df = dfs.get_df_at(caplet.end_time)
        sigma_sqrT = vol * math.sqrt(caplet.end_time)
        d1 = (math.log(
                fwd_rate / caplet.strike) + .5 * sigma_sqrT**2
        ) / sigma_sqrT
        d2 = d1 - sigma_sqrT
        black = fwd_rate * norm.cdf(d1) - caplet.strike * norm.cdf(d2)
        return caplet.notional * yf * df * black

value_caplet(test_cap[1], df_curve, .1)

def value_cap(
    cap: Cap,
    dfs: DiscountFactorCurve,
    vol: float
) -> float:
    return sum([value_caplet(caplet, dfs, vol) for caplet in cap])

Se construyen todos los caps.

all_caps = []
for i in range(1, len(data)):
    all_caps.append(
        make_cap(
            strike=data.iloc[i].tasa_swap,
            num_caplets=i + 1,
            periodicity=.25,
            notional=10000000
        )
    )

Se visualiza un Cap en particular.

which_cap = 2
all_caps[which_cap]

Se valoriza cada Cap y se muestra el resultado. La valorización utiliza la volatilidad entregada en la data.

for which_cap, cap in enumerate(all_caps):
    print(f'Valor de mercado del cap {which_cap + 1}-esimo: \
    {value_cap(cap, df_curve, data.iloc[which_cap + 1].cap_vol):,.0f}')

## Funciones de Valorización HW

def value_caplet_hw(
    caplet: Caplet,
    dfs: DiscountFactorCurve,
    gamma: float,
    sigma: float,
) -> float:
    if caplet.start_time < 0.0:
        return 0.0
    elif caplet.start_time == 0.0:
        yf = caplet.get_yf()
        fwd_rate = dfs.get_linear_fwd_rate_between(
            caplet.start_time,
            caplet.end_time
        )
        return caplet.notional * yf * max(fwd_rate - caplet.strike, 0.0)
    else:
        yf = caplet.get_yf()
        new_strike = 1 / (1 + caplet.strike * yf)
        fwd_rate = dfs.get_linear_fwd_rate_between(
            caplet.start_time,
            caplet.end_time
        )
        new_notional = caplet.notional * (1 + caplet.strike * yf)
        zo = dfs.get_df_at(caplet.start_time)
        zb = dfs.get_df_at(caplet.end_time)
        sz = hw.sz(
            caplet.start_time,
            caplet.end_time,
            gamma,
            sigma
        )
        r0 = dfs.get_continous_rate_at(
            dfs.tenors_dfs[0][0]
        )
        # print(new_strike, zb/zo, r0)
        return new_notional * hw.zcb_call_put(
            hw.CallPut.PUT,
            new_strike,
            caplet.start_time,
            caplet.end_time,
            r0,
            zo,
            zb,
            gamma,
            sigma,
        )

def value_cap_hw(
    cap: Cap,
    dfs: DiscountFactorCurve,
    gamma: float,
    sigma: float,
) -> float:
    return sum([value_caplet_hw(caplet, dfs, gamma, sigma) for caplet in cap])

for which_cap, cap in enumerate(all_caps):
    print(f'Valor de modelo del cap {which_cap + 1}-esimo: \
    {value_cap_hw(cap, df_curve, .5, .005):,.0f}')

def error(gamma_sigma: Iterable[float], *args) -> float:
    caps = args[0]
    dfs = args[1]
    data = args[2]
    result = 0.0
    for i, cap in enumerate(caps):
        result += (value_cap(cap, dfs, data.iloc[i + 1].cap_vol) - 
                   value_cap_hw(cap, dfs, gamma_sigma[0], gamma_sigma[1])) ** 2
    return result

gs0 = [.5, .005]
error(gs0, all_caps, df_curve, data)

## Encuentra `gamma`y `sigma`

res = minimize(
    error,
    gs0,
    args=(all_caps, df_curve, data),
    # method='Powell',
    bounds = [(0.0001, None), (0.0001, None)],
    tol=1e-8
)

res
